% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_2x2_within.R
\name{power_2x2_within}
\alias{power_2x2_within}
\title{Analytic power calculation for 2x2 within designs (needs updating).}
\usage{
power_2x2_within(design_result, alpha_level = 0.05)
}
\arguments{
\item{design_result}{Output from the ANOVA_design function}

\item{alpha_level}{Alpha level used to determine statistical significance}
}
\value{
mu = means

sigma = standard deviation

n = sample size

rho_A = correlation between dependent factors for effect A

rho_B = correlation between dependent factors for effect B

rho_AB = correlation between dependent factors for A*B interaction

alpha_level = alpha level

Cohen_f_A = Cohen's f for main effect A

Cohen_f_B = Cohen's f for main effect B

Cohen_f_AB = Cohen's f for the A*B interaction

lambda_A = lambda for main effect A

lambda_B = lambda for main effect B

lambda_AB = lambda for A*B interaction

critical_F_A = critical F-value for main effect A

critical_F_B = critical F-value for main effect B

critical_F_AB = critical F-value for A*B interaction

power_A = power for main effect A

power_B = power for main effect B

power_AB = power for A*B interaction

mean_mat = matrix of the means
}
\description{
Analytic power calculation for 2x2 within designs (needs updating).
}
\section{References}{

to be added
}

\examples{
design_result <- ANOVA_design(design = "2w*2w", n = 40, mu = c(1, 0, 1, 0),
     sd = 2, r = 0.8, labelnames = c("condition", "cheerful", "sad",
     "voice", "human", "robot"))
power_result <- power_2x2_within(design_result)
}
